requirements.txt
streamlit
pandas
langgraph
langchain
langchain-openai
openai
pydantic
python-dotenv
langchain_community
google.genai
langchain.callbacks